refactor: unify different tools in a single crate
Instead of having multiple distinct tools, do everything in a
busybox-style umbrella tool. While this may seem like an increase of
maintenance burden for the developers, this solution has a number of
advantages for TIR users:
1. Decreased binary sizes: we no longer require a myriad of binaries
   that are almost identical except for the order of function calls. A
   single tool is linked once and contains everythin one would ever need
   from TIR.
2. Easier downstream compiler integration: if a user has a set of custom
   dialects and they want to support all TIR tooling for it, that's now
   as simple as creating a new source file, filling in the correct
   context and calling `tir_main(ctx)`. No need to modify the upstream
   TIR sources or flood your Cargo.toml with boilerplate binaries.

For those who still wish to get a single-purpose tool a set of main
functions is exposed: just pass a context and `None` as arguments,
and you'll get roughly what `tir-opt` or `tir-asm` were before.

As of time of writing thise, there's only two tools available: opt and
asm. However, I beleive it is better to get things straight at the very
beginning, before we run into a refactoring hell.