fix: allow ops with the same name in different dialects
linkme crate does not allow duplicate names in the application. To
workaround this limitation, ask op developers to also differentiate on
dialect name in op definitions as well as trait implementations. While
being a burden on a dialect developer, this allows multiple ops (like
AddOp) to co-exist in different dialects.