feat: functional ISA simulator prototype (#77)
Only a few operations are supported for now. The idea is to give users an ability to set an initial point (registers state and allocated memory) and run a functional mini-simulation. This will not be indicative of the code performance, but allows one to observe the side effects of an assembly snippet. The tool can be used for multiple purposes:
1) Learn assembly by running small examples without the need of writing setup boilerplate
2) Study the behavior of an arbitrary piece of assembly
3) Writing functional tests to see if the generated code is equivalent to the original.

Getting cycle accurate performance estimations is not a priority for this kind of simulation and will not be considered in the near future.