chore: switch to xtask
xtask is a more native alternative to cargo-make. While it's not as
easy to write and understand, it does not require any external setup and
only uses cargo and rustc to get started. Many Rust projects including
cargo and rust-analyzer use xtask idiom as an alternative to Makefiles
and bash scripts.